Fix Android memory mapping (#9), Adreno driver crash (#13), and audit findings
- Fix Android MEM1 mmap failure (fixes #9): In AllocMEM1 on Linux/Android,
  probe candidate addresses and scan 32-bit user space strictly under 4GB
  if 0x80000000 is occupied by Android 11+ ART/zygote mappings.
- Fix Android pipeline compilation crash (fixes #13): Disable background
  pipeline worker thread on Android so shaders are compiled smoothly on
  the main thread via BuildPipelinesPerFrame rather than flooding Qualcomm
  Adreno Vulkan drivers during asset extraction and menu initialization.
- Fix 64-bit pointer truncation in OSRoundUp32B/OSRoundDown32B on TARGET_PC
  by using uintptr_t instead of u32.
- Fix DISC_STRUCT field types and alignment for big-endian disc layouts:
  * itOldottoseaAttributes: change x10 and x28 from s8 to s32, preventing
    modulo/division by zero in HSD_Randi.
  * ScopeBeamAttrs: expand floats array to 10 entries to cover charge
    level 9 without out-of-bounds access on padding.
  * itToolsMotionAttrs: add DISC_STRUCT attribute so nested floats are
    properly byte swapped.
  * ReflectDesc: change x20_behavior from u8 to s32 so reflector flags
    are properly decoded from disc.
  * GroundParam: change x4C_fixed_cam from bool to s32 so Flat Zone fixed
    camera is preserved.
  * itWhiteBeaAttributes: change x14 from s16 to s32.
  * ftKb_DatAttrs: change jumpaerial_unk from s16 to s32.
- Fix command stream endian decoding in grMaterial_801C9490 on PC.

+++ src/melee/lb/types.h
@@ -125,7 +125,7 @@ struct DISC_STRUCT ReflectDesc {
 
     /// @remarks Setting this to 1 causes the reflector to skip ownership
     /// change
-    u8 x20_behavior;
+    s32 x20_behavior;
 };
 DISC_ASSERT_SIZE(struct ReflectDesc, 0x24);
